Addition of Button to CustomPopup
The CustomPopup has been moved to the CustomQT.h file. A button option
has been added to grant additional flexibility with this popup window,
including chaining popups.
