+++ Headers/Main/CustomQT.h
@@ -1,7 +1,12 @@
 #ifndef CUSTOMQT_H
 #define CUSTOMQT_H
 
+#include <qcombobox>
+#include <qlistwidget>
+#include <qtextedit>
 #include <QLabel>
+#include <QDialog>
+#include <QCheckBox>
 
 class CustomLabel : public QLabel {
 public:
@@ -33,6 +38,19 @@ public:
         objectLabel->show();
         return objectLabel;
     };
+};
+
+class CustomPopup : public QDialog{
+public:
+    QCheckBox* checkOption;
+    QComboBox* comboOption;
+    QLineEdit* lineOption;
+    QTextEdit* multiOption;
+    QListWidget* listOption;
+    std::vector<QLineEdit*> boxList;
+    QPushButton* buttonOption;
+
+    void addBlankItem();
 };
 
 #endif // CUSTOMQT_H
